requirements.txt
streamlit==1.31.1
boto3==1.34.34
Pillow==10.2.0
pandas==2.2.0
opencv-python-headless==4.8.1.78
numpy==1.26.3
streamlit-drawable-canvas==0.9.3